+++ src/ft_printf/pf_print_u.c
@@ -18,9 +18,9 @@ static void 	putpad_u(attrib *ph, uiput *db)
 
 int				pf_print_u(attrib *ph, uiput *db)
 {
-	unsigned int	phi;
+	uintmax_t	phi;
 
-	phi = va_arg(db->ap, unsigned int);
+	phi = va_arg(db->ap, uintmax_t);
 	ph->len = ft_numlen(phi, 10);
 	pf_putnbr(phi, ph, db);
 	if (ph->width)
